test(Tests): extract `AssertsSpecEquals`, drop `OA\OpenApi` support
Moved `assertSpecEquals()` out of `OpenApiTestCase` into its own
trait, composed back in. Removed the `OA\OpenApi` auto-`toYaml()`
branch; the three callers that relied on it now call `toYaml()`
explicitly.
