+++ tests/GeneratorTest.php
@@ -36,7 +36,7 @@ final class GeneratorTest extends OpenApiTestCase
             ->setTypeResolver($this->getTypeResolver())
             ->generate($sources);
 
-        $this->assertSpecEquals(file_get_contents(self::getSpecFilename($name)), $openapi);
+        $this->assertSpecEquals(file_get_contents(self::getSpecFilename($name)), $openapi?->toYaml());
     }
 
     public function testScanInvalidSource(): void
